Fix ArgumentOutOfRangeException when loading information from chat files with less than 2 comments

+++ TwitchDownloaderWPF/PageChatUpdate.xaml.cs
@@ -67,7 +67,10 @@ namespace TwitchDownloaderWPF
             try
             {
                 ChatJsonInfo = await ChatJson.DeserializeAsync(InputFile, true, false, CancellationToken.None);
-                ChatJsonInfo.comments.RemoveRange(1, ChatJsonInfo.comments.Count - 2);
+                if (ChatJsonInfo.comments.Count > 2)
+                {
+                    ChatJsonInfo.comments.RemoveRange(1, ChatJsonInfo.comments.Count - 2);
+                }
                 GC.Collect();
             }
             catch (Exception ex)
